fix assertions and add custom config test suite
